VDSEC-126 Fix Sonar warnings and tests

+++ pii-reporting-api/src/main/java/pro/softcom/sentinelle/application/pii/reporting/service/PiiContextExtractor.java
@@ -187,8 +187,8 @@ public class PiiContextExtractor {
 
         // Always include the main entity
         relEntities.add(new TempEntity(
-                Math.clamp(mainStart - lineStartInSource, 0, lineLen),
-                Math.clamp(mainEnd - lineStartInSource, 0, lineLen),
+                Math.clamp((long) mainStart - lineStartInSource, 0, lineLen),
+                Math.clamp((long) mainEnd - lineStartInSource, 0, lineLen),
                 mainType,
                 true
         ));
